move visualization to a separate component
